+++ src/gosb/vtx/README.md
@@ -1,0 +1,3 @@
+# Structure
+
+We are mirroring the gvisor folders layout so we can better re-use their code.
