+++ src/gosb/vtx/cpu.go
@@ -1,4 +1,4 @@
-package kvm
+package vtx
 
 import (
 	"encoding/binary"
