+++ src/gosb/vtx/kernel.go
@@ -1,4 +1,4 @@
-package kvm
+package vtx
 
 import ()
