+++ src/gosb/vtx/vm.go
@@ -1,4 +1,4 @@
-package kvm
+package vtx
 
 /**
 * author: aghosn
