+++ src/gosb/vtx/vm_test.go
@@ -1,4 +1,4 @@
-package kvm
+package vtx
 
 import (
 	"testing"
