+++ src/gosb/vtx/vma.go
@@ -1,4 +1,4 @@
-package kvm
+package vtx
 
 import (
 	gc "gosb/commons"
