+++ src/gosb/vtx/x86.go
@@ -1,4 +1,4 @@
-package kvm
+package vtx
 
 // Segment indices and Selectors.
 const (
